Decoding ATOM_EXT from latin1
By using the encoding crate, decoding ATOM_EXT is simple. Note that we
do not store ATOM_EXT internally, but convert into the same
representation as with ATOM_EXT_UTF8. When sending ATOM_EXT_UTF8 back to
Erlang, Erlang seems to do the reasonable thing and just go with it.